Fix a rendering bug with the x axis
If the x axis wasn't centered vertically, it was drawn as though the
closer to the wrong side e.g. if the range was supposed to be (-2,4),
it was drawn as though the range was (-4,2). Markings etc were correct.

Bug caused by this.center being defined incorrectly (missing negative,
since graph coordinates increase going up, but canvas coordinates do
the opposite).