Fix visual bug in functions with large gradients
When the gradient of a function was too large near the end of the
domain, a gap could be left due to the draw loop not explicitly drawing
to the end of the domain (there is just a for loop with a very small
increment).

Added a context.lineTo() call to explicitly draw that final segment.