Allow Rationals, MathML for <math-plot-text> pos.
Rather than only allowing integers (which was a mistake anyway, should
have been floats) for the `top`, `left`, etc. attributes of the
<math-plot-text> element, Rationals and MathML are now accepted.

Added the MathPlot._parseNumber() function to parse strings which are
either Rationals or MathML elements into floats. This mirrors the
functionality of MathPlot._parseList(), but for individual numbers.

Updated README.