Implement Jest for testing
- Created a Yarn install
- Installed Jest and other dependencies
- Split math-node.js into math-node.js, mathml.js and rational.js, since
  Jest has problems testing any file with WebComponent code (better for
  code clarity anyway).
- Added first test to ensure functionality
- Added testing documentation to README

Note: mathml.js includes some rump code for converting MathML to a
Rational. This isn't properly implemented yet, but was left in because
it was used to test that Jest was working. It has no effect on the
function of the library, and will be implemented properly in the next
commit.