Add RationalTuple class
Previously the class Rational had a static function parseTuple(), which
would return an array of Rationals.

I need to add other functionality involving Rational tuples, so best to
create a class to handle all of them.

The RationalTuple class was added to rational.js, so exports had to be
changed (two classes were now being exported instead of one), so all
classes which imported Rational had to be updated as well.