Allow MathML->Rational conversion to handle lists
The MathML.rational getter can now handle MathML with a <list> element:
it returns a returns a RationalTuple. This is slightly confusing since
there is no longer a guarantee on type of the return, but there
shouldn't ever be any possiblity of confusion.

Added tests to rational.test.js to test creation/behaviour of
RationalTuples (should have been in last commit), and added tests to
mathml.test.js for conversion from MathML to RationalTuples.