Allow labelling of floodfilled regions
Added the `label` attribute to the <math-plot-floodfill> element,
allowing a text label to be added to the filled area.

In order to keep naming consistent with other elements, changed the
`color` attribute to refer to text color, and added a new attribute
`fill-color` for the color of the filled region.

In order to support positioning of the text label, installed the
polylabel module. As this is a Node module and can't be used natively,
used browserify to convert to a frontend-usable format. Added
instructions in README for how to do this again in future if needed, but
should only be necessary if the packages are updated or more Node
modules need to be included.

Added the `node-main.js` and `dependencies.js` files to support these
node-modules. See the Development section of the README for more
details.