Support parsing more MathML elements to Rational
Added support for the following MathML <apply> elements:
    root, sin, cos, tan, abs, ln, log

They were previously supported for converting to functions, but not to
Rationals. For those which return an approximate value, the evaluated
float is passed to the constructor of Rational, which then approximates
it as a rational as it does for any float constructor.

Also added an abs() member function to Rational, which returns a new
Rational, the absolute value of original Rational.

Fixed a bug in the Rational constructor wherein it would fail if it was
passed a float within 1e-9 of an integer, since that, after rounding to
9 decimal places, would become an integer, and the function expected it
to remain a float.

Added tests for all.